solver now has a set boundaries function instead of requireing boundaries on creation. Added new class heightconnecter, which is inherited by everything that needs to react immidiately by a change in heights. Moved confinement and diffusion to this class. Added new sosboundary class which has solver access. Implemented new boundary averageheightboundary which sets the height at the boundary equal to the average within a certain cutoff range. Boundaries now operate in 3D with double precision. Added 1D boundary class for conveinience. Added concentration profile diffusion class which sets a constant concentration for each site (not equal for all sites). No confinement no longer dumps confining surface to file. Fixes to lammpswriter submodule.

+++ src/kmcsolver/boundary/edge.cpp
@@ -3,7 +3,7 @@
 using namespace kMC;
 
 
-Edge::Edge(const int location,
+Edge::Edge(const double location,
            const orientations orientation) :
     FiniteSize(orientation),
     m_location(location)
